Dockerfile: add libguestfs-tools package
In some of our Python tests we build disk images. Currently this involves
using the sudo command for losetup and mount. For some tests sudo is
not necessary if the command guestmount of package libguestfs-tools is
available. Another useful tool in libguestfs-tools is virt-make-fs.

Generally we should get rid of using a root user for building and testing.
As a first step install libguestfs-tools in our Docker image.

Signed-off-by: Heinrich Schuchardt <xypron.glpk@gmx.de>